+++ crates/azure/src/lib.rs
@@ -1,4 +1,5 @@
 mod accounts;
+mod app_service_list_request;
 mod application_gateway_backend_health_request;
 mod application_gateway_list_request;
 mod auth;
@@ -83,6 +84,7 @@ mod virtual_machine_skus;
 mod virtual_machines;
 mod virtual_network;
 pub use crate::accounts::*;
+pub use crate::app_service_list_request::*;
 pub use crate::application_gateway_backend_health_request::*;
 pub use crate::application_gateway_list_request::*;
 pub use crate::auth::*;