+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - Add `ct outage investigate` command
 - Add `ct az network-interface list|show` command
 - Add `ct az private-endpoint list|show` command
+- Add `ct az app-service list|show` command
 
 # v0.34.0